+++ CHANGELOG.md
@@ -5,6 +5,12 @@ see the commit logs. The format is based on
 [Keep a Changelog](https://keepachangelog.com), and this project adheres to
 [Semantic Versioning](https://semver.org).
 
+## Unreleased
+
+### Added
+
+- Facts subcommand.
+
 ## 0.10.1 - 2026-03-11
 
 ### Added